refactor(SenderLink): move sendMessage to SenderLink
 - fix bug in TransferFrame for accepting a channel in options
 - move all sendMessage related code to SenderLink
 - clean up SenderLink.prototype._sendMessage
