+++ lib/frames/transfer_frame.js
@@ -186,8 +186,7 @@ var debug = require('debug')('amqp10:framing:transfer'),
  * @constructor
  */
 function TransferFrame(options) {
-  TransferFrame.super_.call(this);
-  this.channel = 0;
+  TransferFrame.super_.call(this, options.channel);
   if (options instanceof DescribedType) {
     this.readPerformative(options);
     return;
