+++ lib/link.js
@@ -119,18 +119,6 @@ Link.prototype.detach = function() {
   return detachPromise;
 };
 
-Link.prototype.isSender = function() { return this.role === constants.linkRole.sender; };
-
-Link.prototype.canSend = function() {
-  var sendable = (this.linkCredit >= 1 && (!this.session.policy.enableSessionFlowControl || this.session._sessionParams.remoteIncomingWindow >= 1));
-  debug('canSend(' + this.linkCredit + ',' + this.session._sessionParams.remoteIncomingWindow + ') = ' + sendable);
-  return sendable;
-};
-
-Link.prototype.sendMessage = function(message, options) {
-  return this.session.sendMessage(this, message, options);
-};
-
 Link.prototype.addCredits = function(credits, flowOptions) {
   if (this.role === constants.linkRole.sender) {
     throw new errors.InvalidStateError('Cannot add link credits as a sender');
@@ -235,23 +223,6 @@ Link.prototype._messageReceived = function(transferFrame) {
 
   this.emit(Link.MessageReceived, transferFrame.message);
   this._checkCredit();
-};
-
-Link.prototype._sendMessage = function(messageId, message, transferOptions) {
-  if (this.linkCredit <= 0) {
-    throw new errors.OverCapacityError('Cannot send if no link credit.');
-  }
-
-  var opts = transferOptions || {};
-  opts.handle = this.handle;
-  opts.deliveryId = messageId;
-  opts.settled = this.session._sessionParams.senderSettleMode === constants.senderSettleMode.settled;
-  var transferFrame = new TransferFrame(opts);
-  transferFrame.channel = this.session.channel;
-  transferFrame.message = message;
-  this.linkCredit--;
-  this.session.connection.sendFrame(transferFrame);
-  return messageId;
 };
 
 Link.prototype._detachReceived = function(frame) {
